Move login page to /login
