+++ Client/package.json
@@ -18,7 +18,8 @@
     "react-big-calendar": "^1.11.3",
     "react-bootstrap": "^2.10.2",
     "react-dom": "^18.2.0",
-    "react-modal": "^3.16.1"
+    "react-modal": "^3.16.1",
+    "react-router-dom": "^6.22.3"
   },
   "devDependencies": {
     "@types/react": "^18.2.66",
