--- Client/src/App.css
@@ -1,14 +1,0 @@
-#root {
-  text-align: center;
-  background-color: #0f3b2d;
-  height: 100vh;
-  width: 100%;
-}
-
-.Calendar {
-  background-color: whitesmoke;
-}
-
-.card-divs {
-  font-family: Inter, sans-serif;
-}
