+++ Client/src/components/Login.jsx
@@ -1,12 +1,14 @@
 import React, { useState } from "react";
-import logo from './UAB.png';
-export const Login = (props) => {
+import logo from '/uab-color-logo.png';
+import './Login.css';
+export const Login = ({onLogin}) => {
     const [email, setEmail] = useState('');
     const [pass, setPass] = useState('');
 
     const handleSubmit = (e) => {
         e.preventDefault();
         console.log(email);
+        onLogin();
     }
 
 
@@ -16,8 +18,8 @@ return (
         <img src={logo} alt="Logo" className="logo" />
       </div>
       <div className="form-container">
-        <h2>UAB Central Authentication System</h2>
-        <h3>Enter your BlazerID and Password:</h3>
+        <p className="header1">UAB Central Authentication System</p>
+        <p className="header2">Enter your BlazerID and Password:</p>
         <form className="login-form" onSubmit={handleSubmit}>
           <div className="field-container">
             <label htmlFor="blazerid">Blazer ID</label>
@@ -29,15 +31,15 @@ return (
           </div>
           <div className="field-container">
             <button type="submit">Log In</button>
-            <p class="p1">Having trouble logging in?</p>
-            <p class="p2">Contact AskIT at 205-996-5555 if you have any problems using this system or your BlazerID</p>
-            <p class="p3">Security Notice</p>
-            <p class="p4">For security reasons, quit your web browser when finished accessing services that require authentication.
+            <p className="p1">Having trouble logging in?</p>
+            <p className="p2">Contact AskIT at 205-996-5555 if you have any problems using this system or your BlazerID</p>
+            <p className="p3">Security Notice</p>
+            <p className="p4">For security reasons, quit your web browser when finished accessing services that require authentication.
             Be wary of any program or web page that asks for your BlazerID and password. Legitimate UAB web pages promptingfor your BlazerID and
             password should have addresses that begin with https:// (not http:) and contain "uab.edu".Also, your browser should visually indicate
             that you are accessing a secure page, usually a padlock icon. </p>
 
-            <p class="p5">This system is available only for authorized purposes by authorized users.Use for any other purpose may result in disciplinary action
+            <p className="p5">This system is available only for authorized purposes by authorized users.Use for any other purpose may result in disciplinary action
                or criminal prosecution against the user.</p>
           </div>
         </form>
